refactor(ledger): collocate the account-type route registrars :hammer:
Both account-type registrars move out of accounttype_handler.go into
accounttype_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores.

No route, authz tuple, operation ID or path changed.
